♻️ refactor:  routerとthemeをpresentationに移動
Issue #278
