+++ lib/presentation/pages/repo/components/search_repos_query.dart
@@ -5,13 +5,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../config/router.dart';
 import '../../../../domain/repositories/query_history/entities/query_history_input.dart';
 import '../../../../domain/repositories/query_history/query_history_repository.dart';
 import '../../../../localizations/strings.g.dart';
 import '../../../../utils/env/env.dart';
 import '../../../../utils/env/env_define.dart';
 import '../../../../utils/logger.dart';
+import '../../../components/router.dart';
 
 /// リポジトリ検索文字列初期値プロバイダー
 final searchReposInitQueryProvider = Provider<String>(
