+++ lib/presentation/pages/repo/repo_index_page.dart
@@ -4,8 +4,8 @@
 
 import 'package:flutter/material.dart';
 
-import '../../../config/router.dart';
 import '../../../utils/logger.dart';
+import '../../components/router.dart';
 import '../../components/search_app_bar.dart';
 import 'components/repo_list_view.dart';
 import 'components/search_repos_query.dart';
